refactor: `ChristmasPromotionPrecondition`의 패키징 변경
- 구현된 전략을 가지고 있으므로 context 로 분류합니다.
